Move guest image singleton support to core
